+++ tests/data/nauru/population/configs/settings.yaml
@@ -1,0 +1,152 @@
+####################################################################
+# PopulationSim Properties
+####################################################################
+
+
+# Algorithm/Software Configuration
+# ------------------------------------------------------------------
+INTEGERIZE_WITH_BACKSTOPPED_CONTROLS: True
+SUB_BALANCE_WITH_FLOAT_SEED_WEIGHTS: False
+GROUP_BY_INCIDENCE_SIGNATURE: True
+USE_SIMUL_INTEGERIZER: True
+USE_CVXPY: False
+max_expansion_factor: 100000
+num_processes: 1 
+
+
+# Geographic Settings
+# ------------------------------------------------------------------
+geographies: [REGION, PUMA, TAZ]
+seed_geography: PUMA
+
+
+# Tracing
+# ------------------------------------------------------------------
+trace_geography:
+  TAZ: 1
+  #TRACT: 10200
+
+  
+# Data Directory
+# ------------------------------------------------------------------
+data_dir: data
+
+
+# Input Data Tables
+# ------------------------------------------------------------------
+# input_pre_processor input_table_list
+input_table_list:
+  - tablename: households
+    filename : seed_households.csv
+    index_col: hh_id
+    rename_columns:
+      # not sure what SERIALNO is, but it is not unique
+      #SERIALNO: hh_id
+      hhnum: hh_id
+    drop_columns:
+      - SERIALNO
+  - tablename: persons
+    # expression_filename: seed_persons_expressions.csv
+    filename : seed_persons.csv
+    rename_columns:
+      #SERIALNO: hh_id
+      hhnum: hh_id
+      SPORDER: per_num
+    # drop unused mixed type fields to avoid PyTables pipeline performance issues
+    # (PyTables will pickle object types that it cannot map directly to c-types)
+    drop_columns:
+      - indp02
+      - naicsp02
+      - occp02
+      - socp00
+      - occp10
+      - socp10
+      - indp07
+      - naicsp07
+      - SERIALNO
+      - SOCP
+      - NAICSP
+  - tablename: geo_cross_walk
+    filename : geo_cross_walk.csv
+    column_map:
+      TRACTCE: TRACT
+  - tablename: TAZ_control_data
+    filename : control_totals_taz.csv
+  - tablename: REGION_control_data
+    filename : control_totals_meta.csv
+    #filename : scaled_control_totals_meta.csv
+
+
+# Reserved Column Names
+# ------------------------------------------------------------------
+household_weight_col: WGTP
+household_id_col: hh_id
+total_hh_control: num_hh
+
+
+# Control Specification File Name
+# ------------------------------------------------------------------
+control_file_name: controls.csv
+
+
+# Output Tables
+# ------------------------------------------------------------------
+# output_tables can specify either a list of output tables to include or to skip
+# if neither is specified, then no tables will be written
+
+output_tables:
+  action: include
+  tables:
+    - summary_TAZ
+    - summary_TAZ_PUMA
+    - expanded_household_ids
+
+
+# Synthetic Population Output Specification
+# ------------------------------------------------------------------
+#
+
+output_synthetic_population:
+  household_id: hh_id
+  households:
+    filename: synthetic_households.csv
+    columns:
+      - NP
+      - HINCP
+      - WKEXREL
+      - TEN
+      - NR
+      - OCPIP
+
+  persons:
+    filename: synthetic_persons.csv
+    columns:
+      - per_num
+      - AGEP
+      - SEX
+      - OCCP
+      - PWGTP
+      - COW
+      - JWTRNS
+      - RAC1P
+      - RAC2P
+      - RAC3P
+
+
+# Model steps for base mode
+# ------------------------------------------------------------------
+models:
+    - input_pre_processor
+    - setup_data_structures
+    - initial_seed_balancing
+    - meta_control_factoring
+    - final_seed_balancing
+    - integerize_final_seed_weights
+    - sub_balancing.geography=TAZ
+    - expand_households
+    - write_data_dictionary
+    - summarize
+    - write_tables
+    - write_synthetic_population
+
+resume_after:
