+++ tests/data/nauru/population/data/geo_cross_walk.csv
@@ -1,0 +1,22 @@
+"TAZ","PUMA","REGION"
+1,1,1
+2,1,1
+3,1,1
+4,1,1
+5,1,1
+6,1,1
+7,1,1
+8,1,1
+9,1,1
+10,1,1
+11,1,1
+12,1,1
+13,1,1
+14,1,1
+15,1,1
+16,1,1
+17,1,1
+18,1,1
+19,1,1
+20,1,1
+21,1,1
