+++ tests/data/nauru/population/output/final_summary_TAZ.csv
@@ -1,0 +1,21 @@
+geography,id,num_hh_control,female_1_control,female_2_control,female_3_control,female_4_control,female_5_control,female_6_control,female_7_control,female_8_control,female_9_control,female_10_control,female_11_control,female_12_control,female_13_control,female_14_control,female_15_control,female_16_control,female_17_control,female_18_control,male_1_control,male_2_control,male_3_control,male_4_control,male_5_control,male_6_control,male_7_control,male_8_control,male_9_control,male_10_control,male_11_control,male_12_control,male_13_control,male_14_control,male_15_control,male_16_control,male_17_control,male_18_control,num_hh_result,female_1_result,female_2_result,female_3_result,female_4_result,female_5_result,female_6_result,female_7_result,female_8_result,female_9_result,female_10_result,female_11_result,female_12_result,female_13_result,female_14_result,female_15_result,female_16_result,female_17_result,female_18_result,male_1_result,male_2_result,male_3_result,male_4_result,male_5_result,male_6_result,male_7_result,male_8_result,male_9_result,male_10_result,male_11_result,male_12_result,male_13_result,male_14_result,male_15_result,male_16_result,male_17_result,male_18_result,num_hh_diff,female_1_diff,female_2_diff,female_3_diff,female_4_diff,female_5_diff,female_6_diff,female_7_diff,female_8_diff,female_9_diff,female_10_diff,female_11_diff,female_12_diff,female_13_diff,female_14_diff,female_15_diff,female_16_diff,female_17_diff,female_18_diff,male_1_diff,male_2_diff,male_3_diff,male_4_diff,male_5_diff,male_6_diff,male_7_diff,male_8_diff,male_9_diff,male_10_diff,male_11_diff,male_12_diff,male_13_diff,male_14_diff,male_15_diff,male_16_diff,male_17_diff,male_18_diff
+TAZ,2,122,4,15,20,18,16,13,11,16,15,10,9,8,7,5,4,2,0,0,3,13,16,13,12,13,13,17,15,11,8,7,5,3,2,1,0,0,122,4,16,19,18,18,12,12,16,14,10,10,8,7,5,4,2,0,0,3,12,17,14,11,12,13,20,18,11,8,8,5,3,2,1,0,0,0,0,1,-1,0,2,-1,1,0,-1,0,1,0,0,0,0,0,0,0,0,-1,1,1,-1,-1,0,3,3,0,0,1,0,0,0,0,0,0
+TAZ,3,58,2,7,10,9,8,6,5,8,7,5,4,4,3,2,2,1,0,0,1,6,8,6,6,6,6,8,7,5,4,3,2,1,1,0,0,0,58,2,8,10,9,9,6,5,8,6,5,5,5,3,2,2,1,0,0,1,5,8,7,6,6,6,9,9,5,4,3,2,1,1,0,0,0,0,0,1,0,0,1,0,0,0,-1,0,1,1,0,0,0,0,0,0,0,-1,0,1,0,0,0,1,2,0,0,0,0,0,0,0,0,0
+TAZ,4,184,6,23,30,27,25,19,17,24,22,15,13,12,11,8,6,3,0,0,5,19,24,20,18,19,19,26,23,17,12,11,8,5,3,1,0,0,184,6,22,30,27,26,18,18,24,23,15,14,12,11,7,6,3,0,0,5,19,23,19,17,18,19,28,28,17,12,12,8,5,3,1,0,0,0,0,-1,0,0,1,-1,1,0,1,0,1,0,0,-1,0,0,0,0,0,0,-1,-1,-1,-1,0,2,5,0,0,1,0,0,0,0,0,0
+TAZ,5,243,7,31,39,35,33,25,23,32,29,20,17,16,15,11,8,4,1,0,6,25,32,26,23,26,25,34,30,22,16,14,11,6,5,1,0,0,243,7,30,38,34,34,24,24,31,28,20,18,16,15,11,8,4,1,0,6,24,31,25,22,26,26,38,35,22,16,15,11,6,5,1,0,0,0,0,-1,-1,-1,1,-1,1,-1,-1,0,1,0,0,0,0,0,0,0,0,-1,-1,-1,-1,0,1,4,5,0,0,1,0,0,0,0,0,0
+TAZ,6,15,0,2,3,3,2,2,1,2,2,1,1,1,1,0,0,0,0,0,0,2,2,2,2,2,2,2,2,1,1,1,0,0,0,0,0,0,15,0,2,3,2,2,2,2,2,2,1,1,1,1,0,0,0,0,0,0,2,2,1,2,2,3,2,3,1,1,1,0,0,0,0,0,0,0,0,0,0,-1,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,-1,0,0,1,0,1,0,0,0,0,0,0,0,0,0
+TAZ,7,364,11,46,59,53,49,38,34,48,43,30,26,24,22,17,13,5,1,1,9,38,48,38,35,38,38,50,45,33,24,22,16,9,7,2,0,0,364,11,45,58,53,51,36,36,47,42,31,27,24,22,17,13,5,1,1,9,36,47,38,34,38,39,56,54,34,24,23,16,9,7,2,0,0,0,0,-1,-1,0,2,-2,2,-1,-1,1,1,0,0,0,0,0,0,0,0,-2,-1,0,-1,0,1,6,9,1,0,1,0,0,0,0,0,0
+TAZ,8,185,6,24,30,27,25,19,17,25,22,15,13,12,11,9,6,3,0,0,5,19,25,20,18,20,19,26,23,17,12,11,8,5,3,1,0,0,185,6,23,29,27,25,18,19,24,22,15,14,12,11,9,6,3,0,0,5,18,24,19,17,20,19,29,27,17,12,12,8,5,3,1,0,0,0,0,-1,-1,0,0,-1,2,-1,0,0,1,0,0,0,0,0,0,0,0,-1,-1,-1,-1,0,0,3,4,0,0,1,0,0,0,0,0,0
+TAZ,9,547,17,69,88,79,73,57,50,71,65,45,39,35,33,25,19,8,2,1,14,57,71,58,53,57,57,75,67,49,36,33,24,14,11,4,1,1,547,16,67,86,79,75,53,53,69,64,46,41,36,34,25,19,8,2,1,13,54,68,56,50,59,58,83,80,50,37,34,24,14,11,4,0,0,0,-1,-2,-2,0,2,-4,3,-2,-1,1,2,1,1,0,0,0,0,0,-1,-3,-3,-2,-3,2,1,8,13,1,1,1,0,0,0,0,-1,-1
+TAZ,10,149,4,19,24,22,20,15,14,20,18,12,11,9,9,7,5,2,0,0,4,15,20,16,14,16,15,21,18,13,10,9,6,4,3,1,0,0,149,4,19,24,22,21,14,15,20,17,12,12,9,9,7,5,2,0,0,4,13,20,16,14,16,15,23,21,13,10,10,6,4,3,1,0,0,0,0,0,0,0,1,-1,1,0,-1,0,1,0,0,0,0,0,0,0,0,-2,0,0,0,0,0,2,3,0,0,1,0,0,0,0,0,0
+TAZ,11,9,0,1,2,2,2,1,1,2,1,1,1,1,0,0,0,0,0,0,0,1,2,1,1,1,1,2,1,1,1,0,0,0,0,0,0,0,9,0,1,2,2,2,1,2,2,1,1,1,1,0,0,0,0,0,0,0,2,2,1,1,0,1,2,2,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,-1,0,0,1,0,0,0,0,0,0,0,0,0
+TAZ,12,203,6,26,33,30,28,21,19,27,24,17,15,13,12,9,7,3,0,0,5,21,27,22,20,22,21,28,25,18,13,12,9,5,4,1,0,0,203,6,25,33,30,28,20,21,26,23,17,15,12,13,9,7,3,0,0,5,20,26,21,19,23,21,31,30,18,13,12,9,5,4,1,0,0,0,0,-1,0,0,0,-1,2,-1,-1,0,0,-1,1,0,0,0,0,0,0,-1,-1,-1,-1,1,0,3,5,0,0,0,0,0,0,0,0,0
+TAZ,13,26,0,3,4,4,4,3,2,4,3,2,2,2,1,1,1,0,0,0,0,3,4,3,2,3,3,4,3,2,2,1,1,0,0,0,0,0,26,0,2,4,4,4,3,3,3,3,2,2,3,1,1,1,0,0,0,0,4,3,2,2,4,3,4,4,2,2,1,1,0,0,0,0,0,0,0,-1,0,0,0,0,1,-1,0,0,0,1,0,0,0,0,0,0,0,1,-1,-1,0,1,0,0,1,0,0,0,0,0,0,0,0,0
+TAZ,14,380,12,48,61,55,51,40,35,50,45,32,27,24,23,18,13,6,1,1,10,40,50,40,37,40,39,52,46,34,25,23,17,10,8,3,0,0,380,11,46,60,55,53,37,37,49,44,32,28,24,23,18,13,6,1,1,9,37,48,39,35,40,40,58,55,35,25,24,17,10,8,3,0,0,0,-1,-2,-1,0,2,-3,2,-1,-1,0,1,0,0,0,0,0,0,0,-1,-3,-2,-1,-2,0,1,6,9,1,0,1,0,0,0,0,0,0
+TAZ,15,150,4,19,24,22,20,16,14,20,18,12,11,10,9,7,5,2,0,0,4,16,20,16,14,16,16,21,18,13,10,9,6,4,3,1,0,0,150,4,17,24,22,20,15,15,19,18,12,11,10,9,7,5,2,0,0,4,16,18,14,14,15,16,22,23,13,10,10,6,4,3,1,0,0,0,0,-2,0,0,0,-1,1,-1,0,0,0,0,0,0,0,0,0,0,0,0,-2,-2,0,-1,0,1,5,0,0,1,0,0,0,0,0,0
+TAZ,16,45,1,6,8,7,6,5,4,6,5,4,3,3,3,2,1,0,0,0,1,5,6,5,4,5,5,6,6,4,3,3,2,1,1,0,0,0,45,1,7,7,7,8,5,5,5,5,4,4,2,3,2,1,0,0,0,1,6,7,6,4,6,5,8,7,4,3,2,2,1,1,0,0,0,0,0,1,-1,0,2,0,1,-1,0,0,1,-1,0,0,0,0,0,0,0,1,1,1,0,1,0,2,1,0,0,-1,0,0,0,0,0,0
+TAZ,17,175,5,22,29,26,24,18,16,23,21,15,13,11,11,8,6,2,0,0,4,18,23,19,17,18,18,24,22,16,12,10,8,4,3,1,0,0,175,5,21,29,25,25,17,18,22,21,15,13,11,11,8,6,2,0,0,4,18,22,18,16,19,18,26,26,16,12,11,8,4,3,1,0,0,0,0,-1,0,-1,1,-1,2,-1,0,0,0,0,0,0,0,0,0,0,0,0,-1,-1,-1,1,0,2,4,0,0,1,0,0,0,0,0,0
+TAZ,18,104,3,13,17,15,14,11,10,14,12,9,7,7,6,5,3,1,0,0,2,11,14,11,10,11,11,15,13,9,7,6,4,2,2,0,0,0,104,3,13,17,15,15,11,11,14,11,10,7,8,6,5,3,1,0,0,2,10,14,11,10,10,11,17,15,10,7,6,4,2,2,0,0,0,0,0,0,0,0,1,0,1,0,-1,1,0,1,0,0,0,0,0,0,0,-1,0,0,0,-1,0,2,2,1,0,0,0,0,0,0,0,0
+TAZ,19,284,9,36,46,41,38,30,26,37,34,24,20,18,17,13,10,4,1,1,7,30,37,30,27,30,29,39,35,26,19,17,12,7,6,2,0,0,284,9,34,45,41,39,28,28,36,33,24,21,18,17,13,10,4,1,1,7,28,36,29,26,31,30,43,42,26,19,18,12,7,6,2,0,0,0,0,-2,-1,0,1,-2,2,-1,-1,0,1,0,0,0,0,0,0,0,0,-2,-1,-1,-1,1,1,4,7,0,0,1,0,0,0,0,0,0
+TAZ,20,146,4,19,24,21,20,15,14,19,18,12,10,9,9,7,5,2,0,0,4,15,19,16,14,15,15,20,18,13,10,9,6,4,3,1,0,0,146,4,18,24,20,20,14,16,18,17,12,11,9,9,7,5,2,0,0,4,14,19,15,13,15,15,22,21,13,10,10,6,4,3,1,0,0,0,0,-1,0,-1,0,-1,2,-1,-1,0,1,0,0,0,0,0,0,0,0,-1,0,-1,-1,0,0,2,3,0,0,1,0,0,0,0,0,0
+TAZ,21,218,7,27,35,32,29,23,20,29,26,18,16,14,13,10,7,3,0,0,5,23,29,23,21,23,23,30,27,20,14,13,10,5,4,1,0,0,218,7,25,34,32,30,22,21,28,26,18,16,14,14,10,7,3,0,0,5,21,28,22,20,23,23,33,32,20,14,14,10,5,4,1,0,0,0,0,-2,-1,0,1,-1,1,-1,0,0,0,0,1,0,0,0,0,0,0,-2,-1,-1,-1,0,0,3,5,0,0,1,0,0,0,0,0,0
