+++ tests/data/nauru/population/output/final_summary_TAZ_PUMA.csv
@@ -1,0 +1,2 @@
+geography,id,num_hh_control,hh_size_1_control,hh_size_2_3_control,hh_size_4_5_control,hh_size_6_plus_control,female_1_control,female_2_control,female_3_control,female_4_control,female_5_control,female_6_control,female_7_control,female_8_control,female_9_control,female_10_control,female_11_control,female_12_control,female_13_control,female_14_control,female_15_control,female_16_control,female_17_control,female_18_control,male_1_control,male_2_control,male_3_control,male_4_control,male_5_control,male_6_control,male_7_control,male_8_control,male_9_control,male_10_control,male_11_control,male_12_control,male_13_control,male_14_control,male_15_control,male_16_control,male_17_control,male_18_control,num_hh_result,hh_size_1_result,hh_size_2_3_result,hh_size_4_5_result,hh_size_6_plus_result,female_1_result,female_2_result,female_3_result,female_4_result,female_5_result,female_6_result,female_7_result,female_8_result,female_9_result,female_10_result,female_11_result,female_12_result,female_13_result,female_14_result,female_15_result,female_16_result,female_17_result,female_18_result,male_1_result,male_2_result,male_3_result,male_4_result,male_5_result,male_6_result,male_7_result,male_8_result,male_9_result,male_10_result,male_11_result,male_12_result,male_13_result,male_14_result,male_15_result,male_16_result,male_17_result,male_18_result,num_hh_diff,hh_size_1_diff,hh_size_2_3_diff,hh_size_4_5_diff,hh_size_6_plus_diff,female_1_diff,female_2_diff,female_3_diff,female_4_diff,female_5_diff,female_6_diff,female_7_diff,female_8_diff,female_9_diff,female_10_diff,female_11_diff,female_12_diff,female_13_diff,female_14_diff,female_15_diff,female_16_diff,female_17_diff,female_18_diff,male_1_diff,male_2_diff,male_3_diff,male_4_diff,male_5_diff,male_6_diff,male_7_diff,male_8_diff,male_9_diff,male_10_diff,male_11_diff,male_12_diff,male_13_diff,male_14_diff,male_15_diff,male_16_diff,male_17_diff,male_18_diff
+PUMA,1,3607,822,1822,798,165,108,456,586,528,487,377,333,477,430,299,258,233,216,164,121,51,6,4,89,377,477,385,348,381,375,500,444,324,239,214,155,89,69,21,1,1,3607,1039,1666,569,333,106,441,576,524,505,356,361,463,420,302,271,235,219,163,121,51,6,4,87,359,463,373,333,383,381,554,532,328,240,226,155,89,69,21,0,0,0,217,-156,-229,168,-2,-15,-10,-4,18,-21,28,-14,-10,3,13,2,3,-1,0,0,0,0,-2,-18,-14,-12,-15,2,6,54,88,4,1,12,0,0,0,0,-1,-1
