--- tests/data/nauru/population_replace/configs/controls.csv
@@ -1,42 +1,0 @@
-target,geography,seed_table,importance,control_field,expression
-num_hh,TAZ,households,10000000,HHBASE,(households.WGTP > 0) & (households.WGTP < np.inf)
-hh_size_1,REGION,households,5000,HHSIZE1,households.NP == 1
-hh_size_2_3,REGION,households,5000,HHSIZE2,(households.NP == 2) | (households.NP == 3)
-hh_size_4_5,REGION,households,5000,HHSIZE4,(households.NP == 4) | (households.NP == 5)
-hh_size_6_plus,REGION,households,5000,HHSIZE6,households.NP >= 6
-female_1,TAZ,persons,100000,POPF1,(persons.AGEP == 0) & (persons.SEX == 2)
-female_2,TAZ,persons,100000,POPF2,(persons.AGEP >= 1) & (persons.AGEP < 5) & (persons.SEX == 2)
-female_3,TAZ,persons,100000,POPF3,(persons.AGEP >= 5) & (persons.AGEP < 10) & (persons.SEX == 2)
-female_4,TAZ,persons,100000,POPF4,(persons.AGEP >= 10) & (persons.AGEP < 15) & (persons.SEX == 2)
-female_5,TAZ,persons,100000,POPF5,(persons.AGEP >= 15) & (persons.AGEP < 20) & (persons.SEX == 2)
-female_6,TAZ,persons,100000,POPF6,(persons.AGEP >= 20) & (persons.AGEP < 25) & (persons.SEX == 2)
-female_7,TAZ,persons,100000,POPF7,(persons.AGEP >= 25) & (persons.AGEP < 30) & (persons.SEX == 2)
-female_8,TAZ,persons,100000,POPF8,(persons.AGEP >= 30) & (persons.AGEP < 35) & (persons.SEX == 2)
-female_9,TAZ,persons,100000,POPF9,(persons.AGEP >= 35) & (persons.AGEP < 40) & (persons.SEX == 2)
-female_10,TAZ,persons,100000,POPF10,(persons.AGEP >= 40) & (persons.AGEP < 45) & (persons.SEX == 2)
-female_11,TAZ,persons,100000,POPF11,(persons.AGEP >= 45) & (persons.AGEP < 50) & (persons.SEX == 2)
-female_12,TAZ,persons,100000,POPF12,(persons.AGEP >= 50) & (persons.AGEP < 55) & (persons.SEX == 2)
-female_13,TAZ,persons,100000,POPF13,(persons.AGEP >= 55) & (persons.AGEP < 60) & (persons.SEX == 2)
-female_14,TAZ,persons,100000,POPF14,(persons.AGEP >= 60) & (persons.AGEP < 65) & (persons.SEX == 2)
-female_15,TAZ,persons,100000,POPF15,(persons.AGEP >= 65) & (persons.AGEP < 70) & (persons.SEX == 2)
-female_16,TAZ,persons,100000,POPF16,(persons.AGEP >= 70) & (persons.AGEP < 75) & (persons.SEX == 2)
-female_17,TAZ,persons,100000,POPF17,(persons.AGEP >= 75) & (persons.AGEP < 80) & (persons.SEX == 2)
-female_18,TAZ,persons,100000,POPF18,(persons.AGEP >= 80) & (persons.SEX == 2)
-male_1,TAZ,persons,100000,POPM1,(persons.AGEP == 0) & (persons.SEX == 1)
-male_2,TAZ,persons,100000,POPM2,(persons.AGEP >= 1) & (persons.AGEP < 5) & (persons.SEX == 1)
-male_3,TAZ,persons,100000,POPM3,(persons.AGEP >= 5) & (persons.AGEP < 10) & (persons.SEX == 1)
-male_4,TAZ,persons,100000,POPM4,(persons.AGEP >= 10) & (persons.AGEP < 15) & (persons.SEX == 1)
-male_5,TAZ,persons,100000,POPM5,(persons.AGEP >= 15) & (persons.AGEP < 20) & (persons.SEX == 1)
-male_6,TAZ,persons,100000,POPM6,(persons.AGEP >= 20) & (persons.AGEP < 25) & (persons.SEX == 1)
-male_7,TAZ,persons,100000,POPM7,(persons.AGEP >= 25) & (persons.AGEP < 30) & (persons.SEX == 1)
-male_8,TAZ,persons,100000,POPM8,(persons.AGEP >= 30) & (persons.AGEP < 35) & (persons.SEX == 1)
-male_9,TAZ,persons,100000,POPM9,(persons.AGEP >= 35) & (persons.AGEP < 40) & (persons.SEX == 1)
-male_10,TAZ,persons,100000,POPM10,(persons.AGEP >= 40) & (persons.AGEP < 45) & (persons.SEX == 1)
-male_11,TAZ,persons,100000,POPM11,(persons.AGEP >= 45) & (persons.AGEP < 50) & (persons.SEX == 1)
-male_12,TAZ,persons,100000,POPM12,(persons.AGEP >= 50) & (persons.AGEP < 55) & (persons.SEX == 1)
-male_13,TAZ,persons,100000,POPM13,(persons.AGEP >= 55) & (persons.AGEP < 60) & (persons.SEX == 1)
-male_14,TAZ,persons,100000,POPM14,(persons.AGEP >= 60) & (persons.AGEP < 65) & (persons.SEX == 1)
-male_15,TAZ,persons,100000,POPM15,(persons.AGEP >= 65) & (persons.AGEP < 70) & (persons.SEX == 1)
-male_16,TAZ,persons,100000,POPM16,(persons.AGEP >= 70) & (persons.AGEP < 75) & (persons.SEX == 1)
-male_17,TAZ,persons,100000,POPM17,(persons.AGEP >= 75) & (persons.AGEP < 80) & (persons.SEX == 1)
-male_18,TAZ,persons,100000,POPM18,(persons.AGEP >= 80) & (persons.SEX == 1)
