--- tests/data/nauru/population_replace/data/control_totals_meta.csv
@@ -1,2 +1,0 @@
-PUMA,REGION,POPBASE,HHBASE,HHSIZE1,HHSIZE2,HHSIZE4,HHSIZE6
-1,1,9623,3607,822,1822,798,165
