--- tests/data/nauru/population_replace/data/control_totals_taz.csv
@@ -1,21 +1,0 @@
-TAZ,POPF1,POPF2,POPF3,POPF4,POPF5,POPF6,POPF7,POPF8,POPF9,POPF10,POPF11,POPF12,POPF13,POPF14,POPF15,POPF16,POPF17,POPF18,POPM1,POPM2,POPM3,POPM4,POPM5,POPM6,POPM7,POPM8,POPM9,POPM10,POPM11,POPM12,POPM13,POPM14,POPM15,POPM16,POPM17,POPM18,POPBASE,HHBASE1,HHBASE2,HHBASE4,HHBASE6,HHBASE,REGION,MAZ,xTAZ
-2,4,15,20,18,16,13,11,16,15,10,9,8,7,5,4,2,0,0,3,13,16,13,12,13,13,17,15,11,8,7,5,3,2,1,0,0,325,28,61,27,6,122,1,2,2
-3,2,7,10,9,8,6,5,8,7,5,4,4,3,2,2,1,0,0,1,6,8,6,6,6,6,8,7,5,4,3,2,1,1,0,0,0,153,13,29,13,3,58,1,3,3
-4,6,23,30,27,25,19,17,24,22,15,13,12,11,8,6,3,0,0,5,19,24,20,18,19,19,26,23,17,12,11,8,5,3,1,0,0,491,42,93,41,8,184,1,4,4
-5,7,31,39,35,33,25,23,32,29,20,17,16,15,11,8,4,1,0,6,25,32,26,23,26,25,34,30,22,16,14,11,6,5,1,0,0,648,55,123,54,11,243,1,5,5
-6,0,2,3,3,2,2,1,2,2,1,1,1,1,0,0,0,0,0,0,2,2,2,2,2,2,2,2,1,1,1,0,0,0,0,0,0,40,3,8,3,1,15,1,6,6
-7,11,46,59,53,49,38,34,48,43,30,26,24,22,17,13,5,1,1,9,38,48,38,35,38,38,50,45,33,24,22,16,9,7,2,0,0,972,83,184,80,17,364,1,7,7
-8,6,24,30,27,25,19,17,25,22,15,13,12,11,9,6,3,0,0,5,19,25,20,18,20,19,26,23,17,12,11,8,5,3,1,0,0,496,42,94,41,8,185,1,8,8
-9,17,69,88,79,73,57,50,71,65,45,39,35,33,25,19,8,2,1,14,57,71,58,53,57,57,75,67,49,36,33,24,14,11,4,1,1,1458,125,276,121,25,547,1,9,9
-10,4,19,24,22,20,15,14,20,18,12,11,9,9,7,5,2,0,0,4,15,20,16,14,16,15,21,18,13,10,9,6,4,3,1,0,0,396,34,75,33,7,149,1,10,10
-11,0,1,2,2,2,1,1,2,1,1,1,1,0,0,0,0,0,0,0,1,2,1,1,1,1,2,1,1,1,0,0,0,0,0,0,0,27,2,5,2,0,9,1,11,11
-12,6,26,33,30,28,21,19,27,24,17,15,13,12,9,7,3,0,0,5,21,27,22,20,22,21,28,25,18,13,12,9,5,4,1,0,0,543,46,103,45,9,203,1,12,12
-13,0,3,4,4,4,3,2,4,3,2,2,2,1,1,1,0,0,0,0,3,4,3,2,3,3,4,3,2,2,1,1,0,0,0,0,0,67,6,13,6,1,26,1,13,13
-14,12,48,61,55,51,40,35,50,45,32,27,24,23,18,13,6,1,1,10,40,50,40,37,40,39,52,46,34,25,23,17,10,8,3,0,0,1016,87,192,84,17,380,1,14,14
-15,4,19,24,22,20,16,14,20,18,12,11,10,9,7,5,2,0,0,4,16,20,16,14,16,16,21,18,13,10,9,6,4,3,1,0,0,400,34,76,33,7,150,1,15,15
-16,1,6,8,7,6,5,4,6,5,4,3,3,3,2,1,0,0,0,1,5,6,5,4,5,5,6,6,4,3,3,2,1,1,0,0,0,121,10,23,10,2,45,1,16,16
-17,5,22,29,26,24,18,16,23,21,15,13,11,11,8,6,2,0,0,4,18,23,19,17,18,18,24,22,16,12,10,8,4,3,1,0,0,467,40,88,39,8,175,1,17,17
-18,3,13,17,15,14,11,10,14,12,9,7,7,6,5,3,1,0,0,2,11,14,11,10,11,11,15,13,9,7,6,4,2,2,0,0,0,275,24,52,23,5,104,1,18,18
-19,9,36,46,41,38,30,26,37,34,24,20,18,17,13,10,4,1,1,7,30,37,30,27,30,29,39,35,26,19,17,12,7,6,2,0,0,758,65,143,63,13,284,1,19,19
-20,4,19,24,21,20,15,14,19,18,12,10,9,9,7,5,2,0,0,4,15,19,16,14,15,15,20,18,13,10,9,6,4,3,1,0,0,390,33,74,32,7,146,1,20,20
-21,7,27,35,32,29,23,20,29,26,18,16,14,13,10,7,3,0,0,5,23,29,23,21,23,23,30,27,20,14,13,10,5,4,1,0,0,580,50,110,48,10,218,1,21,21
