--- tests/data/nauru/population_replace/output/mem.csv
@@ -1,40 +1,0 @@
-process,pid,rss,full_rss,uss,event,children,time
-MainProcess,28002,141_774_848,141_774_848,95_535_104,pipeline.run before preload_injectables,0,2022/08/31 20:24:02.036064
-MainProcess,28002,141_774_848,141_774_848,95_535_104,pipeline.run after preload_injectables,0,2022/08/31 20:24:02.039675
-MainProcess,28002,141_774_848,141_774_848,95_535_104,pipeline.run_model input_pre_processor start,0,2022/08/31 20:24:02.044513
-MainProcess,28002,144_814_080,144_814_080,98_369_536,pipeline.run_model input_pre_processor finished,0,2022/08/31 20:24:02.136441
-MainProcess,28002,149_131_264,149_131_264,102_707_200,pipeline.run after input_pre_processor,0,2022/08/31 20:24:02.206251
-MainProcess,28002,149_131_264,149_131_264,102_707_200,pipeline.run_model setup_data_structures start,0,2022/08/31 20:24:02.214256
-MainProcess,28002,151_216_128,151_216_128,104_050_688,pipeline.run_model setup_data_structures finished,0,2022/08/31 20:24:02.374500
-MainProcess,28002,153_440_256,153_440_256,106_221_568,pipeline.run after setup_data_structures,0,2022/08/31 20:24:02.477508
-MainProcess,28002,153_440_256,153_440_256,106_221_568,pipeline.run_model initial_seed_balancing start,0,2022/08/31 20:24:02.482686
-MainProcess,28002,153_440_256,153_440_256,106_287_104,pipeline.run_model initial_seed_balancing finished,0,2022/08/31 20:24:03.613840
-MainProcess,28002,153_440_256,153_440_256,106_287_104,pipeline.run after initial_seed_balancing,0,2022/08/31 20:24:03.642448
-MainProcess,28002,153_440_256,153_440_256,106_287_104,pipeline.run_model meta_control_factoring start,0,2022/08/31 20:24:03.648530
-MainProcess,28002,153_440_256,153_440_256,106_287_104,pipeline.run_model meta_control_factoring finished,0,2022/08/31 20:24:03.664568
-MainProcess,28002,153_440_256,153_440_256,106_287_104,pipeline.run after meta_control_factoring,0,2022/08/31 20:24:03.692195
-MainProcess,28002,153_440_256,153_440_256,106_287_104,pipeline.run_model final_seed_balancing start,0,2022/08/31 20:24:03.697573
-MainProcess,28002,153_440_256,153_440_256,106_287_104,pipeline.run_model final_seed_balancing finished,0,2022/08/31 20:24:04.969612
-MainProcess,28002,153_440_256,153_440_256,106_287_104,pipeline.run after final_seed_balancing,0,2022/08/31 20:24:04.996156
-MainProcess,28002,153_440_256,153_440_256,106_287_104,pipeline.run_model integerize_final_seed_weights start,0,2022/08/31 20:24:05.002943
-MainProcess,28002,165_081_088,165_081_088,118_300_672,pipeline.run_model integerize_final_seed_weights finished,0,2022/08/31 20:24:05.079974
-MainProcess,28002,165_081_088,165_081_088,118_304_768,pipeline.run after integerize_final_seed_weights,0,2022/08/31 20:24:05.107647
-MainProcess,28002,165_081_088,165_081_088,118_304_768,pipeline.run_model sub_balancing.geography=TAZ start,0,2022/08/31 20:24:05.113390
-MainProcess,28002,169_361_408,169_361_408,121_966_592,pipeline.run_model sub_balancing.geography=TAZ finished,0,2022/08/31 20:24:22.018524
-MainProcess,28002,170_168_320,170_168_320,122_945_536,pipeline.run after sub_balancing.geography=TAZ,0,2022/08/31 20:24:22.064438
-MainProcess,28002,170_168_320,170_168_320,122_945_536,pipeline.run_model expand_households start,0,2022/08/31 20:24:22.070142
-MainProcess,28002,170_168_320,170_168_320,123_195_392,pipeline.run_model expand_households finished,0,2022/08/31 20:24:22.153433
-MainProcess,28002,170_168_320,170_168_320,123_195_392,pipeline.run after expand_households,0,2022/08/31 20:24:22.181905
-MainProcess,28002,170_168_320,170_168_320,123_195_392,pipeline.run_model write_data_dictionary start,0,2022/08/31 20:24:22.187134
-MainProcess,28002,169_000_960,169_000_960,121_704_448,pipeline.run_model write_data_dictionary finished,0,2022/08/31 20:24:22.340658
-MainProcess,28002,171_163_648,171_163_648,123_801_600,pipeline.run after write_data_dictionary,0,2022/08/31 20:24:22.362442
-MainProcess,28002,171_163_648,171_163_648,123_801_600,pipeline.run_model summarize start,0,2022/08/31 20:24:22.368017
-MainProcess,28002,171_163_648,171_163_648,123_801_600,pipeline.run_model summarize finished,0,2022/08/31 20:24:22.517015
-MainProcess,28002,170_082_304,170_082_304,122_753_024,pipeline.run after summarize,0,2022/08/31 20:24:22.595971
-MainProcess,28002,170_082_304,170_082_304,122_753_024,pipeline.run_model write_tables start,0,2022/08/31 20:24:22.601794
-MainProcess,28002,170_082_304,170_082_304,122_753_024,pipeline.run_model write_tables finished,0,2022/08/31 20:24:22.617248
-MainProcess,28002,170_082_304,170_082_304,122_753_024,pipeline.run after write_tables,0,2022/08/31 20:24:22.637913
-MainProcess,28002,170_082_304,170_082_304,122_753_024,pipeline.run_model write_synthetic_population start,0,2022/08/31 20:24:22.643750
-MainProcess,28002,172_072_960,172_072_960,124_829_696,pipeline.run_model write_synthetic_population finished,0,2022/08/31 20:24:22.697675
-MainProcess,28002,172_072_960,172_072_960,124_829_696,pipeline.run after write_synthetic_population,0,2022/08/31 20:24:22.718212
-MainProcess,28002,172_072_960,172_072_960,124_829_696,pipeline.run after run_models,0,2022/08/31 20:24:22.723764
