--- tests/data/nauru/population_replace/output/timing_log.csv
@@ -1,13 +1,0 @@
-process_name,model_name,seconds,minutes
-MainProcess,input_pre_processor,0.2,0.0
-MainProcess,setup_data_structures,0.3,0.0
-MainProcess,initial_seed_balancing,1.2,0.0
-MainProcess,meta_control_factoring,0.0,0.0
-MainProcess,final_seed_balancing,1.3,0.0
-MainProcess,integerize_final_seed_weights,0.1,0.0
-MainProcess,sub_balancing.geography=TAZ,17.0,0.3
-MainProcess,expand_households,0.1,0.0
-MainProcess,write_data_dictionary,0.2,0.0
-MainProcess,summarize,0.2,0.0
-MainProcess,write_tables,0.0,0.0
-MainProcess,write_synthetic_population,0.1,0.0
