+++ tradesman/model_creation/synthetic_population/seeds_url.py
@@ -1,0 +1,1 @@
+population_url = "https://github.com/AequilibraE/tradesman/releases/download/V0.1b/population.zip"
